#568 worked out how to check if qos handler present and kill Ocelot if options specified but no handler, need to refactor this into fluent validation style
