+++ src/Ocelot/Configuration/Validator/FileConfigurationFluentValidator.cs
@@ -15,11 +15,12 @@ namespace Ocelot.Configuration.Validator
 
     public class FileConfigurationFluentValidator : AbstractValidator<FileConfiguration>, IConfigurationValidator
     {
-        private readonly IServiceProvider _provider;
+        private readonly QosDelegatingHandlerDelegate _qosDelegatingHandlerDelegate;
 
         public FileConfigurationFluentValidator(IAuthenticationSchemeProvider authenticationSchemeProvider, IServiceProvider provider)
         {
-            _provider = provider;
+            _qosDelegatingHandlerDelegate = provider.GetService<QosDelegatingHandlerDelegate>();
+
             RuleFor(configuration => configuration.ReRoutes)
                 .SetCollectionValidator(new ReRouteFluentValidator(authenticationSchemeProvider, provider));
 
@@ -44,15 +45,13 @@ namespace Ocelot.Configuration.Validator
                 .WithMessage((config, aggregateReRoute) => $"{nameof(aggregateReRoute)} {aggregateReRoute.UpstreamPathTemplate} contains ReRoute with specific RequestIdKey, this is not possible with Aggregates");
         }
 
-        private bool AllReRoutesForAggregateExist(FileAggregateReRoute fileAggregateReRoute, List<FileReRoute> reRoutes)
-        {
-            var reRoutesForAggregate = reRoutes.Where(r => fileAggregateReRoute.ReRouteKeys.Contains(r.Key));
-
-            return reRoutesForAggregate.Count() == fileAggregateReRoute.ReRouteKeys.Count;
-        }
-
         public async Task<Response<ConfigurationValidationResult>> IsValid(FileConfiguration configuration)
         {
+            if (HasQos(configuration) && _qosDelegatingHandlerDelegate == null)
+            {
+                return new OkResponse<ConfigurationValidationResult>(new ConfigurationValidationResult(true, new FileValidationFailedError("Unable to start Ocelot because either a ReRoute or GlobalConfiguration are using QoSOptions but no QosDelegatingHandlerDelegate has been registered in dependency injection container.")));
+            }
+
             var validateResult = await ValidateAsync(configuration);
 
             if (validateResult.IsValid)
@@ -67,7 +66,19 @@ namespace Ocelot.Configuration.Validator
             return new OkResponse<ConfigurationValidationResult>(result);
         }
 
-        private static bool DoesNotContainReRoutesWithSpecificRequestIdKeys(FileAggregateReRoute fileAggregateReRoute, 
+        private bool HasQos(FileConfiguration configuration)
+        {
+            return configuration.ReRoutes.Any(x => x.QoSOptions.TimeoutValue > 0 && x.QoSOptions.ExceptionsAllowedBeforeBreaking > 0) || (configuration.GlobalConfiguration.QoSOptions.TimeoutValue > 0 && configuration.GlobalConfiguration.QoSOptions.ExceptionsAllowedBeforeBreaking > 0);
+        }
+
+        private bool AllReRoutesForAggregateExist(FileAggregateReRoute fileAggregateReRoute, List<FileReRoute> reRoutes)
+        {
+            var reRoutesForAggregate = reRoutes.Where(r => fileAggregateReRoute.ReRouteKeys.Contains(r.Key));
+
+            return reRoutesForAggregate.Count() == fileAggregateReRoute.ReRouteKeys.Count;
+        }
+
+        private static bool DoesNotContainReRoutesWithSpecificRequestIdKeys(FileAggregateReRoute fileAggregateReRoute,
             List<FileReRoute> reRoutes)
         {
             var reRoutesForAggregate = reRoutes.Where(r => fileAggregateReRoute.ReRouteKeys.Contains(r.Key));
@@ -75,15 +86,15 @@ namespace Ocelot.Configuration.Validator
             return reRoutesForAggregate.All(r => string.IsNullOrEmpty(r.RequestIdKey));
         }
 
-        private static bool IsNotDuplicateIn(FileReRoute reRoute, 
+        private static bool IsNotDuplicateIn(FileReRoute reRoute,
             List<FileReRoute> reRoutes)
         {
             var matchingReRoutes = reRoutes
-                .Where(r => r.UpstreamPathTemplate == reRoute.UpstreamPathTemplate 
+                .Where(r => r.UpstreamPathTemplate == reRoute.UpstreamPathTemplate
                             && (r.UpstreamHost != reRoute.UpstreamHost || reRoute.UpstreamHost == null))
                 .ToList();
 
-            if(matchingReRoutes.Count == 1)
+            if (matchingReRoutes.Count == 1)
             {
                 return true;
             }
